+++ corehq/apps/cleanup/tests/__init__.py
@@ -1,1 +1,1 @@
-from test_fix_forms_with_missing_xmlns import *
+from test_fix_forms_and_apps_with_missing_xmlns import *
